init: cross platform image builder without docker
We recently added support for building cross platform docker containers using docker buildx. However, docker does cross platform builds by emulating the target arch using qemu. That is obviously very slow to build and consumes 100% of the cpu, and when run on arm64 machines has a tendency to fail completely.

This in contrast, is a small go binary which uses `GOOS= GOARCH= go build` directly to compile the binaries and then uses `google/go-containerregistry` to create and publish the new containers based on the desired platform.

This is inspired by `google/ko` which is awesome but doesn't support multiple binaries in a single container.

Signed-off-by: Maisem Ali <maisem@tailscale.com>