Allow suppressing build output
Currently mkctr runs all builds with `go build -v`. The verbose
build logs are a measurable fraction of build times in CI. This
patch allows people to opt-out of the build logs. It is intended
that interested consumers should configure their pipelines to set
`-v` if they are debugging the build.