mkctr.go: add --annotations flag
Adds a new --annotations flag that can be used to provide OCI annotations as key/value pairs.
https://github.com/opencontainers/image-spec/blob/main/annotations.md
Annotations must be provided as a string array of key=value pairs, i.e key1=val1,key2=val2.
Annotations with empty values are not supported.
For single image manifests, annotations are added to the manifest.
For manifest lists (image indexes), annotations are added to each image manifest
as well as the image index.

Updates tailscale/tailscale#12914

Signed-off-by: Irbe Krumina <irbe@tailscale.com>